refactor(workflow): reorganize imports and add execution ID utility
